requirements.txt
streamlit
python-dotenv
google-genai
language-tool-python
gtts
speechrecognition